# Conflicts:
#	api/db/crud.py
#	api/routes/mri.py

+++ api/routes/mri.py
@@ -159,8 +159,13 @@ async def remove_annotation(
     )
 
     try:
-        await crud.delete_annotation(annotation_id)
-        service.files().delete(fileId=annotation.file_id).execute()
+        if annotation.is_ai is True:
+            # soft delete
+            update_values = { "visible": False }
+            await crud.update_annotation_details(annotation_id, update_values)
+        else:
+            await crud.delete_annotation(annotation_id)
+            service.files().delete(fileId=annotation.file_id).execute()
     except Exception:
         raise APIException(
             status_code=status.HTTP_404_NOT_FOUND,
@@ -198,6 +203,39 @@ async def change_annotation(
             content={"message": translation["annotation_name_exists"]},
         )
     return Response(status_code=status.HTTP_204_NO_CONTENT)
+
+
+@router.patch(
+    "/{mri_id}/ai_annotation",
+    response_model=s.Annotation,
+    dependencies=[Depends(validate_api_token)]
+)
+async def request_ai_annotation_visibility(
+    mri_id: int,
+    user_id: int = Depends(validate_api_token),
+    translation=Depends(get_localization_data)
+):
+    mri = await utils.verify_file_creator(
+        mri_id,
+        user_id,
+        "mri",
+        translation
+    )
+    try:
+        annotation = await crud.get_ai_annotation_by_mri_id(mri.id)
+        # tu moze nastat viacero pripadov, kedy by sme mohli spustat inferenciu:
+        # 1. v DB nie je zaznam o AI anotacii
+        # 2. v DB je, ale nema file a job je ''
+        # mozno este 3. ma file, ale neda sa najst na Drive, lebo bol zmazany
+        update_values = { "visible": True }
+        await crud.update_annotation_details(annotation.id, update_values)
+
+    except Exception:
+        raise APIException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            content={"message": translation["annotation_not_found"]},
+        )
+    return annotation
 
 
 @router.post('/annotations/ai')
